backend/pyproject.toml
[project]
name = "agent-cypher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "neo4j==5.14.1",
  "retry==0.9.2",
  "python-dotenv==1.1.1",
  "loguru==0.7.3",

  "redis==6.4.0",
  "langchain==0.2.17",
  "langchain-openai==0.1.25",
  "google-generativeai==0.7.2",
  "langchain-google-genai==1.0.8",
  "openai==1.109.1",
  "langchain-community==0.2.19",
  "langchain-core==0.2.43",
  "opentelemetry-api==1.22.0",
  "pydantic==2.7.4",
  "uvicorn==0.25.0",
  "asyncio==3.4.3",
  "fastapi==0.109.0",
  "numpy==1.26.2"
]
